新增了注释；删除了Vuex中的chatReader部分
对重构后的代码进行了注释；同时合并了chatReader和Chat，方便实现管理，合并之后命名为chat
