+++ src/components/chatHeadDu.vue
@@ -28,8 +28,9 @@
       likesDu
     },  
     setup(){
+      // 集中管理了留言信息的显示
       const store = useStore()
-      const chatInfo = computed(()=> store.state.chatReader.chatInfo)
+      const chatInfo = computed(()=> store.state.chats.chatInfo)
       
       return{
         chatInfo
