+++ src/pages/commuDu.vue
@@ -38,6 +38,7 @@
 
           }
         })
+        // 获取数据之后处理过滤数据
         const filteredArticlesByTag = computed(() => store.getters['articles/filteredArticlesByTag']('生活'));
 
         // 响应式设计相关
